requirements.txt
fastapi
uvicorn[standard]
groq
httpx
playwright
pdfplumber
Pillow
google.generativeai